+++ src/components/Home.vue
@@ -1,17 +1,17 @@
 <template>
+  <div style="height: 100%">
   <div class="personal-center" style="background-color: #eceff1;height: 100%">
-    <el-row type="flex"   justify="center" align="middle" class="row-bg" style="height: 250px;flex-direction: column" >
+    <el-row type="flex"   justify="center" align="middle" class="row-bg" style="height: 250px;flex-direction: row;margin-left: 50px;" >
       <el-col :span="8" style="width: 170px;height: 200px" >
-        <el-avatar :size="200" :src="user.avatar" @error="errorHandler" ></el-avatar>
+        <el-avatar :size="200" :src="getAvatar()" ></el-avatar>
       </el-col>
-      <el-col :span="8" style="margin-left: 35px;margin-top: 10px; width: 100px; ">
-        <h1>{{ user.name }}</h1>
-      </el-col>
+      <el-button  type="info" icon="el-icon-edit" style="width: 2%;height: 10%;margin-left: 10px;padding: 0;margin-top: 180px;"  size="small"  @click="changeUserImage()">
+      </el-button>
     </el-row>
     <el-row type="flex" justify="center" align="middle" class="row-bg" style="min-height: 300px;flex-direction: column">
-<!--      <el-col :span="18" style="margin-left: 35px;margin-top: 10px; width: 200px; ">
-        <h1>{{'名字：'+ user.name }}</h1>
-      </el-col>-->
+      <el-col :span="8" style="margin-left: 35px;margin-top: 10px; width: 100px;font-size: 30px; ">
+        <h1>{{ user.name }}</h1>
+      </el-col>
       <el-col :span="18" style="margin-left: 50px;margin-top: 50px; width: 200px; ">
         <h1>{{'账号：'+ user.number }}</h1>
       </el-col>
@@ -19,12 +19,36 @@
         <h1>{{'电话：'+ user.phone }}</h1>
       </el-col>
       <el-col :span="18" style="margin-left: 50px;margin-top: 50px; width: 200px; ">
-        <h1>{{ this.user.role === 0 ? '权限：超级管理员' : (scope.row.roleId === 1 ? '权限：管理员' : '权限：用户') }}</h1>
+        <h1>{{ this.user.role === 0 ? '权限：超级管理员' :  '权限：用户' }}</h1>
       </el-col>
     </el-row>
-    <el-col :span="24" style="background-color: #eceff1">
+    <el-col :span="24" style="background-color: #eceff1;margin-top: 10px;">
       <el-button @click="logout" type="danger" style="margin-left: 50px" >退出登录</el-button>
     </el-col>
+  </div>
+  <el-dialog
+      title="更改头像"
+      :visible.sync="imagedialogVisible"
+      width="30%">
+    <el-radio-group v-model="radio" >
+      <el-radio :label="1">
+        <el-avatar :size="50" :src="require(`@/assets/images/QCF.png`)" ></el-avatar>
+      </el-radio>
+      <el-radio :label="2">
+        <el-avatar :size="50" :src="require(`@/assets/images/ZJ.png`)" ></el-avatar>
+      </el-radio>
+      <el-radio :label="3">
+        <el-avatar :size="50" :src="require(`@/assets/images/ZS.png`)" ></el-avatar>
+      </el-radio>
+      <el-radio :label="4">
+        <el-avatar :size="50" :src="require(`@/assets/images/GLMX.png`)" ></el-avatar>
+      </el-radio>
+    </el-radio-group>
+    <span slot="footer" class="dialog-footer">
+    <el-button @click="imagedialogVisible = false">取 消</el-button>
+    <el-button type="primary" @click="changeImageHandle">确 定</el-button>
+  </span>
+  </el-dialog>
   </div>
 </template>
 <script>
@@ -32,9 +56,13 @@ export default {
   name: "HoMe",
   data() {
     return {
+      radio: 1,//单选选中选项
+      imagedialogVisible:false,//单选框显示
+      tempImage:'',
+      userId:'',//用户id
+      userImage:'',//头像文件
       user: {
         name: sessionStorage.getItem('curUser').replace(/"/g,''), // 示例用户名
-        avatar: require("@/assets/images/QCF.png") ,// 头像路径
         number: "a",
         phone:"这是电话",
         role:''
@@ -42,6 +70,59 @@ export default {
     };
   },
   methods: {
+    changeUserImage(){
+      this.imagedialogVisible=true;
+    },
+    changeImageHandle(){
+      switch (this.radio) {
+        case 1:
+          this.tempImage = 'QCF'
+          break
+        case 2:
+          this.tempImage = 'ZJ'
+          break
+        case 3:
+          this.tempImage = 'ZS'
+          break
+        case 4:
+          this.tempImage = 'GLMX'
+          break
+        default:
+          this.tempImage = 'YH';
+          break
+      }
+      this.$axios.get(this.$httpUrl + '/users/saveImage',{
+        params: {
+          id: this.userId,
+          image:this.tempImage
+        }}).then(res => res.data).then(
+          res => {
+            if (res.code === 1) {
+              this.$message({
+                type: 'success',
+                message: '修改成功'
+              })
+              console.log(res.data)
+            } else {
+              alert("更改失败")
+            }
+          }
+      )
+      this.imagedialogVisible=false;
+    },
+    getImage(){
+      this.userId= sessionStorage.getItem('curId').replace(/"/g,'')
+      this.$axios.get(this.$httpUrl + '/users/findImageById?id='+this.userId).then(res => res.data).then(
+          res => {
+            if (res.code === 1) {
+              this.userImage=res.data+'.png';
+              sessionStorage.setItem('userImage',this.userImage);
+              console.log(this.userImage)
+            } else {
+              alert("获取数据失败")
+            }
+          })
+    },
     getuser(){
       this.$axios.get(this.$httpUrl + '/users/findByName',{
         params: {
@@ -59,8 +140,8 @@ export default {
           }
       )
     },
-    errorHandler() {
-      return true
+    getAvatar(){
+      return require(`@/assets/images/${this.userImage}`);
     },
     logout() {
       this.$confirm("您确定要退出登录吗？",'提示',{
@@ -69,12 +150,23 @@ export default {
         center:true
       })
           .then(()=>{
-            this.$message({
-              type:'success',
-              message:'您已经退出登录成功'
-            })
-            this.$router.push("/")
-            sessionStorage.clear();
+            this.$axios.get(this.$httpUrl + '/logout?id='+this.userId).then(res => res.data).then(
+                res => {
+                  if (res.code === 1) {
+                    this.$message({
+                      type: 'success',
+                      message: '您已经退出登录'
+                    })
+                    sessionStorage.clear();
+                    this.$router.push("/")
+                  } else {
+                    this.$message({
+                      type: 'danger',
+                      message: '退出登录失败，请稍后再试'
+                    })
+                  }
+                }
+            )
           }).catch(()=>{
         this.$message({
           type:'info',
@@ -83,8 +175,10 @@ export default {
       })
     },
   },
+  //组件实例创建后调用，否则接收不到函数
   beforeMount() {
     this.getuser();
+    this.getImage();
   }
 };
 </script>
